# StreamSyncBackend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: StreamSync API
  version: 1.0.0
  description: Real-time collaborative workspace backend. Workspaces, projects, documents,
    tasks, activity and AI assistance.
paths:
...
  /api/projects/:
    get:
      operationId: projects_list
      summary: List projects
      parameters:
      - in: query
        name: ordering
        schema:
          type: string
          enum:
          - -created_at
          - -due_date
          - -name
          - -updated_at
          - created_at
          - due_date
          - name
          - updated_at
        description: Sort field. Defaults to -updated_at.
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - in: query
        name: search
        schema:
          type: string
        description: Match name or description.
      - in: query
        name: status
        schema:
          type: string
          enum:
          - active
          - archived
          - completed
          - on_hold
          - planning
        description: Filter by lifecycle status.
      - in: query
        name: workspace
        schema:
          type: string
        description: Filter by workspace id.
      tags:
      - projects
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedProjectList'
          description: ''
    post:
      operationId: projects_create
      description: Requires the editor or owner role in the target workspace.
      summary: Create a project
      tags:
      - projects
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreateRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
          description: ''
        '403':
          description: Viewers cannot create projects.
        '404':
          description: No such workspace for this caller.
  /api/projects/{project_id}/:
    get:
      operationId: projects_retrieve
      summary: Project detail
      parameters:
      - in: path
        name: project_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - projects
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
          description: ''
        '404':
          description: Not found.
    patch:
      operationId: projects_update
      description: Requires the editor or owner role. Viewers are read-only.
      summary: Update a project
      parameters:
      - in: path
        name: project_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - projects
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedProjectUpdateRequest'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
          description: ''
        '403':
          description: Read-only role.
    delete:
      operationId: projects_destroy
      description: Workspace owner only. Documents filed under the project survive
        and become unfiled rather than being deleted with it.
      summary: Delete a project
      parameters:
      - in: path
        name: project_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - projects
      security:
      - jwtAuth: []
      responses:
        '204':
          description: Deleted.
...
  /api/search/:
    get:
      operationId: search_list
      description: Scoped to the caller's workspaces. Results are ranked across types
        and capped; queries shorter than two characters return an empty list.
      summary: Search documents, projects, tasks and people
      parameters:
      - in: query
        name: q
        schema:
          type: string
        description: The query. Under two characters returns [].
      - in: query
        name: workspace
        schema:
          type: string
        description: Optional workspace id to narrow the search to one workspace.
      tags:
      - search
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SearchResult'
          description: ''
...
components:
  schemas:
...
    PaginatedProjectList:
      type: object
      required:
      - count
      - page
      - page_size
      - total_pages
      - results
      properties:
        count:
          type: integer
          example: 42
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 25
        total_pages:
          type: integer
          example: 2
        next:
          type: string
          nullable: true
          format: uri
        previous:
          type: string
          nullable: true
          format: uri
        results:
          type: array
          items:
            $ref: '#/components/schemas/Project'
...
    PatchedProjectUpdateRequest:
      type: object
      description: Partial update. Every field optional; workspace and owner are not
        here.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          nullable: true
          maxLength: 2000
        status:
          $ref: '#/components/schemas/ProjectStatus'
        due_date:
          type: string
          format: date
          nullable: true
...
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        workspace_id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          readOnly: true
        slug:
          type: string
          readOnly: true
          description: URL-safe identifier, unique within the workspace.
          pattern: ^[-a-zA-Z0-9_]+$
        description:
          type: string
          readOnly: true
        status:
          allOf:
          - $ref: '#/components/schemas/ProjectStatus'
          readOnly: true
        task_count:
          type: integer
          description: |-
            Annotated by the view's queryset, so this costs no query per project.

            The fallback keeps the serializer usable outside that queryset — a
            correctness path, not the expected one.
          readOnly: true
        completed_task_count:
          type: integer
          readOnly: true
        due_date:
          type: string
          format: date
          readOnly: true
          nullable: true
        members:
          type: array
          items:
            type: object
            additionalProperties: {}
          description: |-
            Who can work on this project.

            There is no project-level membership: access to a project *is* access
            to its workspace (README §7), so the members are the workspace's active
            members. The view prefetches them, so this costs no query per row.
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - completed_task_count
      - created_at
      - description
      - due_date
      - id
      - members
      - name
      - slug
      - status
      - task_count
      - updated_at
      - workspace_id
    ProjectCreateRequest:
      type: object
      description: |-
        Creation input.

        `workspace_id` is in the body because the frontend posts to a flat
        `/api/projects/` rather than a nested route. It is validated against the
        caller's own workspaces in the view — the id itself is never trusted.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          nullable: true
          default: ''
          maxLength: 2000
        status:
          allOf:
          - $ref: '#/components/schemas/ProjectStatus'
          default: planning
        due_date:
          type: string
          format: date
          nullable: true
        workspace_id:
          type: string
          format: uuid
      required:
      - name
      - workspace_id
    ProjectStatus:
      enum:
      - planning
      - active
      - on_hold
      - completed
      - archived
      type: string
      description: |-
        * `planning` - Planning
        * `active` - Active
        * `on_hold` - On hold
        * `completed` - Completed
        * `archived` - Archived
...
    SearchResult:
      type: object
      properties:
        id:
          type: string
          readOnly: true
        type:
          allOf:
          - $ref: '#/components/schemas/SearchResultTypeEnum'
          readOnly: true
        title:
          type: string
          readOnly: true
        subtitle:
          type: string
          readOnly: true
          nullable: true
        href:
          type: string
          readOnly: true
          nullable: true
        score:
          type: number
          format: double
          readOnly: true
      required:
      - href
      - id
      - score
      - subtitle
      - title
      - type
    SearchResultTypeEnum:
      enum:
      - document
      - project
      - task
      - person
      type: string
      description: |-
        * `document` - document
        * `project` - project
        * `task` - task
        * `person` - person
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT